test database works. now to fix the unit tests that were already written for fcq, and to write association tests for course and instructor.
