+++ Gemfile
@@ -65,6 +65,5 @@ group :production do
 end
 
 group :test do
-  gem 'factory_girl_rails'
   gem 'minitest-spec-rails'
 end
