+++ app/assets/stylesheets/application.css
@@ -15,15 +15,15 @@
 
 *=require bootstrap-wysihtml5
 
-.pretty th .current {  
-  padding-right: 12px;  
-  background-repeat: no-repeat;  
-  background-position: right center;  
-}  
-.pretty th .asc {   
-  background-image: url(/images/up_arrow.gif);   
-}  
-  
-.pretty th .desc {   
-  background-image: url(/images/down_arrow.gif);   
-} 
+.pretty th .current {
+  padding-right: 12px;
+  background-repeat: no-repeat;
+  background-position: right center;
+}
+.pretty th .asc {
+  background-image: url(/images/up_arrow.gif);
+}
+
+.pretty th .desc {
+  background-image: url(/images/down_arrow.gif);
+}
