+++ test/test_prepare.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+rake db:test:prepare
+rake RAILS_ENV=test db:seed
