+++ total_reload.sh
@@ -10,8 +10,9 @@ rm -rf solr/development
 rm -rf solr/pids
 rm -rf solr/test
 
+RAILS_ENV=development
 #startup solr in development environment
-rake sunspot:solr:start RAILS_ENV=development
+rake sunspot:solr:start
 
 #all of our rake tasks
 bundle exec rake db:reset
